first-pass commit, expanding combat component into its subcomponents
