+++ services/app/src/components/views/quest/cardtemplates/Render.tsx
@@ -31,3 +31,32 @@ export function generateIconElements(content: string, theme: CardThemeType): JSX
     });
   return <span dangerouslySetInnerHTML={{__html: content }} />;
 }
+
+export const NUMERALS: {[k: number]: string; } = {
+  1: 'I',
+  2: 'II',
+  3: 'III',
+  4: 'IV',
+  5: 'V',
+};
+
+function numberToWord(input: number): string {
+  switch (input) {
+    case 0: return 'zero';
+    case 1: return 'one';
+    case 2: return 'two';
+    case 3: return 'three';
+    case 4: return 'four';
+    case 5: return 'five';
+    case 6: return 'six';
+    case 7: return 'seven';
+    case 8: return 'eight';
+    case 9: return 'nine';
+    case 10: return 'ten';
+    default: return input.toString();
+  }
+}
+
+function capitalizeFirstLetter(input: string): string {
+  return input.charAt(0).toUpperCase() + input.slice(1);
+}
