+++ services/app/src/components/views/quest/cardtemplates/combat/Defeat.test.tsx
@@ -53,58 +53,8 @@ function setup(phase: CardPhase, overrides: Partial<Props>) {
   return {props, enzymeWrapper};
 }
 
-describe('Combat', () => {
-  describe('DRAW_ENEMIES', () => {
-    it('renders all enemies in props', () => {
-      const combat = newCombat(TEST_NODE);
-      const {enzymeWrapper} = setup('DRAW_ENEMIES', {combat});
-      expect(enzymeWrapper.find('h2.draw_enemies').map((e) => e.text())).toEqual([
-        'Thief (Tier I )',
-        'Brigand (Tier I )',
-        'Footpad (Tier I )',
-      ]);
-    });
-  });
-
-  describe('NO_TIMER', () => {
-    it('shows non-timer prep card');
-  });
-
-  describe('PREPARE', () => {
-    it('shows preparation card');
-  });
-
-  describe('TIMER', () => {
-    it('shows timer card');
-  });
-
-  describe('SURGE', () => {
-    it('shows surge card');
-  });
-
-  describe('RESOLVE_ABILITIES', () => {
-    it('shows horror persona helper');
-    it('shows rolls if enabled in settings');
-  });
-
-  describe('RESOLVE_DAMAGE', () => {
-    it('starts at current player and tier count');
-  });
-
-  describe('VICTORY', () => {
-    it('shows a victory page');
-    it('shows healing if not suppressed');
-    it('shows loot if not suppressed');
-    it('shows levelup if not suppressed');
-  });
-
-  describe('DEFEAT', () => {
-    it('does not show Retry button if onLose does not go to **end**');
-    it('shows Retry button if onLose goes to **end**');
-    it('hitting Retry returns the state to the card before combat');
-  });
-
-  describe('MID_COMBAT_ROLEPLAY', () => {
-    it('shows the current parsernode in a dark theme');
-  });
+describe('Combat defeat', () => {
+  it('does not show Retry button if onLose does not go to **end**');
+  it('shows Retry button if onLose goes to **end**');
+  it('hitting Retry returns the state to the card before combat');
 });
