+++ services/app/src/components/views/quest/cardtemplates/combat/DefeatContainer.test.tsx
@@ -1,7 +1,5 @@
-describe('CombatContainer', () => {
+describe('DefeatContainer', () => {
   it('calculates max tier from history');
-
   it('handles undefined combat element');
-
   it('skips the timer card on prev button');
 });
