+++ services/app/src/components/views/quest/cardtemplates/combat/DrawEnemiesContainer.tsx
@@ -1,0 +1,65 @@
+import {toCard, toPrevious} from 'app/actions/Card';
+import {event} from 'app/actions/Quest';
+import {MAX_ADVENTURER_HEALTH} from 'app/Constants';
+import {logEvent} from 'app/Logging';
+import {getMultiplayerClient} from 'app/Multiplayer';
+import {EventParameters} from 'app/reducers/QuestTypes';
+import {AppStateWithHistory, SettingsType} from 'app/reducers/StateTypes';
+import {getStore} from 'app/Store';
+import {connect} from 'react-redux';
+import Redux from 'redux';
+import {
+  midCombatChoice,
+} from '../roleplay/Actions';
+import {ParserNode} from '../TemplateTypes';
+import {
+  adventurerDelta,
+  generateCombatTemplate,
+  handleCombatEnd,
+  handleCombatTimerHold,
+  handleCombatTimerStart,
+  handleCombatTimerStop,
+  handleResolvePhase,
+  setupCombatDecision,
+  tierSumDelta,
+} from './Actions';
+import Combat, {DispatchProps, StateProps} from './Combat';
+import {CombatPhase, CombatState} from './Types';
+
+const mapStateToProps = (state: AppStateWithHistory, ownProps: Partial<StateProps>): StateProps => {
+  const node = ownProps.node;
+  if (!node || !card) {
+    throw Error('Incomplete props given');
+  }
+  const combatFromNode = (node && node.ctx && node.ctx.templates && node.ctx.templates.combat);
+  const combat: CombatState = combatFromNode || generateCombatTemplate(state.settings, false, state.quest.node, getStore().getState);
+  const stateCombat = (state.quest.node && state.quest.node.ctx && state.quest.node.ctx.templates && state.quest.node.ctx.templates.combat)
+    || {tier: 0, mostRecentRolls: [10], numAliveAdventurers: 1};
+
+  // Override with dynamic state for tier
+  // Any change causes a repaint
+  return {
+    node: state.quest.node,
+    combat,
+    settings: state.settings,
+    tier: stateCombat.tier,
+  };
+};
+
+const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
+  return {
+    onNext: (phase: CombatPhase) => {
+      dispatch(toCard({name: 'QUEST_CARD', phase}));
+    },
+    onTierSumDelta: (node: ParserNode, current: number, delta: number) => {
+      dispatch(tierSumDelta({node, current, delta}));
+    },
+  };
+};
+
+const CombatContainer = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Combat);
+
+export default CombatContainer;
