+++ services/app/src/components/views/quest/cardtemplates/combat/MidCombatRoleplay.test.tsx
@@ -1,0 +1,3 @@
+describe('Combat MidCombatRoleplay', () => {
+  it('shows the current parsernode in a dark theme');
+});
