+++ services/app/src/components/views/quest/cardtemplates/combat/NoTimer.test.tsx
@@ -1,0 +1,3 @@
+describe('Combat NoTimer', () => {
+  it('shows non-timer prep card');
+});
