+++ services/app/src/components/views/quest/cardtemplates/combat/PlayerTier.test.tsx
@@ -1,0 +1,3 @@
+describe('Combat PlayerTier', () => {
+  it('starts at current player and tier count');
+});
