+++ services/app/src/components/views/quest/cardtemplates/combat/PrepareTimer.test.tsx
@@ -1,0 +1,3 @@
+describe('Combat PrepareTimer', () => {
+  it('shows preparation card');
+});
