+++ services/app/src/components/views/quest/cardtemplates/combat/Resolve.test.tsx
@@ -1,0 +1,4 @@
+describe('Combat resolve', () => {
+  it('shows horror persona helper');
+  it('shows rolls if enabled in settings');
+});
