+++ services/app/src/components/views/quest/cardtemplates/combat/Resolve.tsx
@@ -1,0 +1,77 @@
+import AudioControlsContainer from 'app/components/base/AudioControlsContainer';
+import Button from 'app/components/base/Button';
+import Callout from 'app/components/base/Callout';
+import Card from 'app/components/base/Card';
+import Picker from 'app/components/base/Picker';
+import TimerCard from 'app/components/base/TimerCard';
+import {MAX_ADVENTURER_HEALTH, NODE_ENV} from 'app/Constants';
+import {Enemy, EventParameters, Loot} from 'app/reducers/QuestTypes';
+import {CardState, MultiplayerState, SettingsType} from 'app/reducers/StateTypes';
+import * as React from 'react';
+import {REGEX} from 'shared/Regex';
+import Roleplay from '../roleplay/Roleplay';
+import {ParserNode} from '../TemplateTypes';
+import {isSurgeNextRound, roundTimeMillis} from './Actions';
+import {CombatPhase, CombatState} from './Types';
+
+export interface StateProps {
+  mostRecentRolls?: number[];
+  settings: SettingsType;
+}
+
+export interface DispatchProps {
+  onNext: (phase: CombatPhase) => void;
+  onReturn: () => void;
+}
+
+export interface Props extends StateProps, DispatchProps {}
+
+export default resolve(props: Props); : JSX.Element; {
+  let helpText: JSX.Element = (<p>Resolve all played abilities.</p>);
+  const theHorror = (props.settings.contentSets.horror === true);
+  if (props.settings.showHelp) {
+    helpText = (
+      <span>
+        {theHorror && <div>
+          <h2>The Horror <img className="inline_icon" src="images/horror_white_small.svg"></img></h2>
+          <p>Adventurers at Min persona must resolve their persona effect and reset to Base persona before resolving abilities. Adventurers at Max persona may choose to resolve and reset now or in a later round.</p>
+        </div>}
+        <h2>Roll <img className="inline_icon" src="images/roll_white_small.svg"></img></h2>
+        <p>
+          Each adventurer rolls a die for each ability they played. If <img className="inline_icon" src="images/roll_white_small.svg"></img> &ge; X, the ability succeeds. Ability cards may list additional effects based on the roll, even if they fail.
+        </p>
+        <h2>Resolve & Discard <img className="inline_icon" src="images/cards_white_small.svg"></img></h2>
+        <p>
+          <strong>Resolve</strong> your abilities in any order. You may change the resolution order during the round, even after some abilities have been played.
+        </p>
+        <p>
+          <strong>Discard</strong> all ability(s) you resolved this round. Put unplayed abilities back into your draw pile before shuffling.
+        </p>
+        <h2>Modifiers <img className="inline_icon" src="images/damage_white_small.svg"></img></h2>
+        <p>
+          Enemies may have modifiers on their cards that affect the damage they receive. The modifier applies to the final damage value of each card played against them.
+        </p>
+
+      </span>
+    );
+  }
+  let renderedRolls: JSX.Element[]|null = null;
+  if (props.settings.autoRoll && props.mostRecentRolls) {
+    renderedRolls = props.mostRecentRolls.map((roll: number, index: number) => {
+      return (<div className="roll" key={index}>{roll}</div>);
+    });
+  }
+
+  return (
+    <Card title="Roll &amp; Resolve" theme="dark" inQuest={true} onReturn={() => props.onReturn()}>
+      {helpText}
+      {renderedRolls &&
+        <div>
+          {props.settings.showHelp && <p>Resolve your abilities with the following rolls. Start with the last person to read the quest and go clockwise:</p>}
+          <div className="rolls">{renderedRolls}</div>
+        </div>
+      }
+      <Button onClick={() => props.onNext('RESOLVE_DAMAGE')}>Next</Button>
+    </Card>
+  );
+}
