+++ services/app/src/components/views/quest/cardtemplates/combat/ResolveContainer.tsx
@@ -1,0 +1,63 @@
+import {toCard, toPrevious} from 'app/actions/Card';
+import {event} from 'app/actions/Quest';
+import {MAX_ADVENTURER_HEALTH} from 'app/Constants';
+import {logEvent} from 'app/Logging';
+import {getMultiplayerClient} from 'app/Multiplayer';
+import {EventParameters} from 'app/reducers/QuestTypes';
+import {AppStateWithHistory, SettingsType} from 'app/reducers/StateTypes';
+import {getStore} from 'app/Store';
+import {connect} from 'react-redux';
+import Redux from 'redux';
+import {
+  midCombatChoice,
+} from '../roleplay/Actions';
+import {ParserNode} from '../TemplateTypes';
+import {
+  adventurerDelta,
+  generateCombatTemplate,
+  handleCombatEnd,
+  handleCombatTimerHold,
+  handleCombatTimerStart,
+  handleCombatTimerStop,
+  handleResolvePhase,
+  setupCombatDecision,
+  tierSumDelta,
+} from './Actions';
+import Resolve, {DispatchProps, StateProps} from './Resolve';
+import {CombatPhase, CombatState} from './Types';
+
+const mapStateToProps = (state: AppStateWithHistory, ownProps: Partial<StateProps>): StateProps => {
+  const node = ownProps.node;
+  if (!node || !card) {
+    throw Error('Incomplete props given');
+  }
+
+  const combatFromNode = (node && node.ctx && node.ctx.templates && node.ctx.templates.combat);
+  const combat: CombatState = combatFromNode || generateCombatTemplate(state.settings, false, state.quest.node, getStore().getState);
+  const stateCombat = (state.quest.node && state.quest.node.ctx && state.quest.node.ctx.templates && state.quest.node.ctx.templates.combat)
+    || {tier: 0, mostRecentRolls: [10], numAliveAdventurers: 1};
+
+  // Override with dynamic state for tier and adventurer count
+  // Any combat param change (e.g. change in tier) causes a repaint
+  return {
+    mostRecentRolls: stateCombat.mostRecentRolls,
+    settings: state.settings,
+  };
+};
+
+const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
+  return {
+    onNext: (phase: CombatPhase) => {
+      dispatch(toCard({name: 'QUEST_CARD', phase}));
+    },
+    onReturn: () => {
+      // Return to the "Ready for Combat?" card instead of doing the timed round again.
+      dispatch(toPrevious({before: false, skip: [{name: 'QUEST_CARD', phase: 'TIMER'}]}));
+    },
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Resolve);
