+++ services/app/src/components/views/quest/cardtemplates/combat/Surge.test.tsx
@@ -1,0 +1,3 @@
+describe('Combat surge', () => {
+  it('shows surge card');
+});
