+++ services/app/src/components/views/quest/cardtemplates/combat/TimerCard.test.tsx
@@ -1,0 +1,3 @@
+describe('Combat TimerCard', () => {
+  it('shows timer card');
+});
