+++ services/app/src/components/views/quest/cardtemplates/combat/Victory.test.tsx
@@ -1,0 +1,6 @@
+describe('Combat victory', () => {
+  it('shows a victory page');
+  it('shows healing if not suppressed');
+  it('shows loot if not suppressed');
+  it('shows levelup if not suppressed');
+});
